+++ src/data-model/syntax-tree/UseSchemaNode.ts
@@ -13,8 +13,14 @@ export class UseSchemaNode extends GenericNode {
   private line: string;
   private useSchemaLength: number = "USE SCHEMA".length;
   private schemaText: JSON | null;
+  private schemaPath: string;
 
-  constructor(lineNumber: number, line: string, schemaText: JSON | null) {
+  constructor(
+    lineNumber: number,
+    line: string,
+    schemaText: JSON | null,
+    schemaPath: string
+  ) {
     super([line], IndexRange.create(lineNumber, 0, lineNumber, line.length), [
       new KeywordNode(
         [line.substring(0, "USE SCHEMA".length)],
@@ -24,7 +30,13 @@ export class UseSchemaNode extends GenericNode {
     this.lineNumber = lineNumber;
     this.line = line;
     this.schemaText = schemaText;
+    this.schemaPath = schemaPath;
   }
+
+  public get filePath() {
+    return "file://" + this.schemaPath;
+  }
+
   public getRelevantChildren(): GenericNode[] {
     const childList: GenericNode[] = [];
     return childList;